Enable each platfrom to configure its own TCR setting
Currently the TCR bits are hardcoded in xlat_tables.c. In order to
map higher physicall address into low virtual address, the TCR bits
need to be configured accordingly. This patch is to enable a platform
specific TCR setting.

The enable_mmu_el is changed to take a TCR parameter, which is passed
from platform dependent code