Juno: Add support for PSCI cpu_suspend api
This patch adds support for the PSCI cpu_suspend api to allow entry
into low power states until affinity level 1 i.e. cluster. It mainly
ensures that a consolidated power off command which specifies the
level to which each affinity level should be powered down is sent to
the SCP.

Change-Id: I84201f5c06e9c41c9e1fbc6438d12ab5f65ee429